[Feat/#63] 지도 마커 커스텀

+++ Offroad-iOS/Offroad-iOS/Network/Place/DTO/Responce/RegisteredPlaceResponseDTO.swift
@@ -22,7 +22,7 @@ struct RegisteredPlaceInfo: Codable {
     let address: String
     let shortIntroduction: String
     let placeCategory: String
-    let latitude: Float
-    let longitude: Float
+    let latitude: Double
+    let longitude: Double
     let visitCount: Int
 }
